--- HolaMundo.java
@@ -1,9 +1,0 @@
-package holamundo;
-
-public class HolaMundo {
-    
-    public static void main(String[] args) {
-        System.out.println("Hola mundo !!!");
-    }
-    
-}
