+++ Sintaxis_JavaEjemplos/src/main/java/HolaMundo.java
@@ -1,0 +1,9 @@
+package holamundo;
+
+public class HolaMundo {
+    
+    public static void main(String[] args) {
+        System.out.println("Hola mundo !!!");
+    }
+    
+}
