+++ Sintaxis_JavaEjemplos/src/main/java/operaAsignacion.java
